+++ ntex-net/src/polling/mod.rs
@@ -4,7 +4,7 @@ mod reactor;
 mod stream;
 
 pub use self::reactor::{Handler, Reactor, ReactorApi};
-pub use ntex_polling::{Event, PollMode};
+pub use ::polling::{Event, PollMode};
 
 #[cfg(not(target_pointer_width = "64"))]
 compile_error!("Only 64bit platforms are supported");